requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.0
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20
python-multipart==0.0.6
cryptography==41.0.7
python-dotenv==1.0.0
